moved configGenerator again
