+++ src/main/scala/org/apache/spark/deploy/armada/submit/ConfigGenerator.scala
@@ -14,17 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.scheduler.cluster.armada
+package org.apache.spark.deploy.armada.submit
 
 import k8s.io.api.core.v1.generated._
 import org.apache.spark.SparkConf
+import org.apache.spark.deploy.armada.submit.ConfigGenerator.REMOTE_CONF_DIR_NAME
 import org.apache.spark.deploy.k8s.Constants.ENV_SPARK_CONF_DIR
 
 import java.io.File
 import scala.io.Source
 
+private[submit] object ConfigGenerator {
+  val REMOTE_CONF_DIR_NAME = "/opt/spark/conf"
+}
 
-class ConfigGenerator(val confDirName: String, val prefix: String,
+private[submit] class ConfigGenerator(val prefix: String,
                       val conf: SparkConf) {
   private val confDir = Option(conf.getenv(ENV_SPARK_CONF_DIR)).orElse(
     conf.getOption("spark.home").map(dir => s"$dir/conf"))
@@ -46,7 +50,7 @@ class ConfigGenerator(val confDirName: String, val prefix: String,
   def getVolumeMounts: Seq[VolumeMount] = {
     Seq(VolumeMount()
       .withName(prefix + "-volume")
-      .withMountPath(confDirName)
+      .withMountPath(REMOTE_CONF_DIR_NAME)
       .withReadOnly(true))
   }
